split agenton compositor modules
